ebb发送trunked成功

+++ ebb/include/net_ebb_response.h
@@ -23,11 +23,11 @@ int net_ebb_response_append_head(net_ebb_response_t response, const char * name,
 int net_ebb_response_append_head_line(net_ebb_response_t response, const char * head_line);
 
 int net_ebb_response_append_body_identity_begin(net_ebb_response_t response, uint32_t size);
-int net_ebb_response_append_body_identity_data(net_ebb_response_t response, void const * data, size_t data_size);
+int net_ebb_response_append_body_identity_data(net_ebb_response_t response, void const * data, uint32_t data_size);
 int net_ebb_response_append_body_identity_from_stream(net_ebb_response_t response, read_stream_t rs);
 
 int net_ebb_response_append_body_chunked_begin(net_ebb_response_t response);
-int net_ebb_response_append_body_chunked_block(net_ebb_response_t response, void const * data, size_t data_size);
+int net_ebb_response_append_body_chunked_block(net_ebb_response_t response, void const * data, uint32_t data_size);
 
 int net_ebb_response_append_complete(net_ebb_response_t response);
